💄 ✨ ⚡️ 🎨 (UI): Added Tweet component also removed the navigation based on client feedback. Also changed mock login credentials to 1 and 1 respectively
