+++ client/src/components/Home/Home.tsx
@@ -1,9 +1,5 @@
 import { useState } from "react";
-
-// importing icons for the sorting options
-import { AiOutlineHeart, AiOutlineRetweet } from "react-icons/ai";
-import { FaRegCommentAlt } from "react-icons/fa";
-// import Tweet from "../Tweet/Tweet";
+import Tweet from "../Tweet/Tweet";
 
 // importing mock data
 import tweeter from "../../mock.json";
@@ -58,63 +54,24 @@ const Home = (props: any) => {
     }
   };
 
-  // style for the icons
-  const style_ = { color: "black", fontSize: "1rem" };
-
   // tweet options
   const tweetOptions = [];
 
   for (let index = 1; index <= 100; index++) {
-    tweetOptions.push(<option key={index}>{index}</option>);
+    tweetOptions.push(<option key={index.toString()}>{index}</option>);
   }
 
   // processing api response
-  const apiResponse = [<div></div>];
-
-  // index = 0
+  const apiResponse = [<div key={"begining div"}></div>];
 
   tweeter.tweets.map(
-    (data, index = 0) =>
+    (data) =>
       data.tags.toLowerCase().match(enteredSearch.toLowerCase()) &&
       enteredSearch !== "" &&
-      apiResponse.push(
-        <div
-          key={index}
-          data-testid="results"
-          className=" w-full border-b border-l border-r border-gray-200 flex flex-col p-3"
-        >
-          {/* <Tweet tweetData={data} num={index} /> */}
-          <div className="flex flex-row items-center">
-            <p className=" font-semibold">{data.name}</p>
-            &nbsp;
-            <p className=" font-bold">&sdot;</p>
-            &nbsp;
-            <p className="text-sm">{data.date}</p>
-          </div>
-
-          <div className=" pt-3 flex flex-row">
-            <p>{data.tweet}</p>
-          </div>
-
-          <div className="flex flex-row justify-around pt-3">
-            <p className="flex flex-row text-sm items-center">
-              <FaRegCommentAlt style={style_} />
-              &nbsp;
-              {parseInt(data.comments, 10)}
-            </p>
-            <p className="flex flex-row text-sm items-center">
-              <AiOutlineRetweet style={style_} />
-              &nbsp; {parseInt(data.retweets, 10)}
-            </p>
-            <p className="flex flex-row text-sm items-center">
-              <AiOutlineHeart style={style_} />
-              &nbsp;
-              {parseInt(data.likes, 10)}
-            </p>
-          </div>
-        </div>
-      )
+      apiResponse.push(<Tweet tweetData={data} />)
   );
+
+  let ind = 0;
 
   return (
     <div data-testid="home">
@@ -149,7 +106,7 @@ const Home = (props: any) => {
       <div className="flex flex-row flex-wrap justify-around pt-3 pb-3 border border-gray-200 items-center">
         {/*  */}
 
-        <div className="flex flex-row w-1/3 justify-center">
+        <div className="flex flex-row flex-wrap w-1/3 justify-center">
           <p>Tweets:</p> &nbsp;
           <select className=" text-black" onChange={tweetHandler}>
             {tweetOptions}
@@ -157,7 +114,7 @@ const Home = (props: any) => {
         </div>
 
         {/* this is for the Fitlering options */}
-        <div className="flex flex-row w-1/3 justify-center">
+        <div className="flex flex-row flex-wrap w-1/3 justify-center">
           <p className="">Filter:</p> &nbsp;
           <select className=" text-black" onChange={filterHandler}>
             <option>by likes</option>
@@ -167,7 +124,7 @@ const Home = (props: any) => {
         </div>
 
         {/* this is for the sorting options */}
-        <div className="flex flex-row w-1/3 justify-center">
+        <div className="flex flex-row flex-wrap w-1/3 justify-center">
           <p className="">Sort:</p> &nbsp;
           <select className=" text-black" onChange={sortHandler}>
             <option>ascending order</option>
@@ -204,7 +161,7 @@ const Home = (props: any) => {
         {apiResponse}
 
         {enteredSearch === "" && clicked === false && (
-          <div className="mt-4">
+          <div className="mt-4" key={(ind++).toString()}>
             <h1 className="text-2xl">Trends</h1>
           </div>
         )}
