+++ client/src/mock.json
@@ -6005,6 +6005,12 @@
       "name": "_go.sh_",
       "email": "go.shoderu@gmail.com",
       "password": "pass"
+    },
+    {
+      "id": 1002,
+      "name": "1",
+      "email": "go.shoderu@gmail.com",
+      "password": "1"
     }
   ],
   "tweets": [
